requirements.txt
asgiref==3.7.2
autoflake==2.2.1
colorama==0.4.6
dj-rest-auth==5.0.2
Django==4.2.6
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
pillow==10.2.0
pyflakes==3.2.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
